Add target availability constraint for SHUFFLE()
This adds an 'excluded cards' parameter to the shuffle function, indicating which portion of the deck should not be included in a shuffle.
When a card says to shuffle the deck (or a part of it), the deck (or that part) needs to have at least one card in it.